+++ src/parsers/mod.rs
@@ -1,0 +1,3 @@
+pub mod code_review;
+pub mod magit;
+pub mod utils;
